Complete rewrite of producer and clearup of tests.
This commit brings in structured data support for RFC5424 and, more importantly,
a much cleaner producer removing a lot of silly code. We also accept all fields
that are possible to be filled in, and default appropriately when they are
empty.

In addition to this, the BSD style tests have been rewritten to be more inline
with the RFC.